requirements.txt
bleak
bluepy
mido
nanoleafapi
numpy
opencv-python
python-dotenv
python-rtmidi